updates are getting saved in notifcation table and getting displaced on ecomm dashboard.

+++ src/main/java/com/acms/CentralSellerPortal/Entities/Product.java
@@ -39,7 +39,7 @@ public class Product implements Serializable {
     public Product() {
     }
 
-    public Product(int product_id, @NotNull String product_name, @NotNull int price, @NotNull String product_description, @NotNull int discount, Seller seller) {
+    public Product(long product_id, @NotNull String product_name, @NotNull int price, @NotNull String product_description, @NotNull int discount, Seller seller) {
         this.productId = product_id;
         this.productName = product_name;
         this.price = price;
